requirements.txt
torch
transformers
scikit-learn
tqdm
pdfplumber
pdf2image
pytesseract
Pillow
easyocr
numpy
 